Fix release clippy gate
(cherry picked from commit ed879a666b22911b336fc0dfc8659d6107872e8c)

+++ crates/kubio-observe/src/observer.rs
@@ -802,8 +802,7 @@ impl Observer {
                 vec![AdaptiveReuseBlocker::ProtectedRoute],
             );
         }
-        if route.shadow_mismatches as u64 > self.adaptive_reuse.public_object.max_shadow_mismatches
-        {
+        if route.shadow_mismatches > self.adaptive_reuse.public_object.max_shadow_mismatches {
             return ReuseEligibility::blocked(
                 ReuseClass::HardProtected,
                 vec![AdaptiveReuseBlocker::ShadowMismatch],
