+++ crates/kubio-observe/src/state.rs
@@ -149,7 +149,7 @@ impl RouteStats {
         config.enabled
             && public_object.enabled
             && self.state != RouteState::Protected
-            && (self.shadow_mismatches as u64) <= public_object.max_shadow_mismatches
+            && self.shadow_mismatches <= public_object.max_shadow_mismatches
             && self.request_count >= public_object.min_route_samples
             && self.distinct_key_count() >= public_object.min_distinct_keys
             && self.dynamic_value_count() >= public_object.min_distinct_keys
@@ -164,7 +164,7 @@ impl RouteStats {
             && self.state != RouteState::Protected
             && self.path_sensitive_samples == 0
             && (self.id_like_path_samples > 0 || self.distinct_key_count() > 1)
-            && (self.shadow_mismatches as u64) <= config.public_object.max_shadow_mismatches
+            && self.shadow_mismatches <= config.public_object.max_shadow_mismatches
     }
 
     pub(crate) fn reuse_class(&self, config: &AdaptiveReuseConfig) -> ReuseClass {
@@ -197,7 +197,7 @@ impl RouteStats {
         if self.state == RouteState::Protected {
             blockers.push(AdaptiveReuseBlocker::ProtectedRoute);
         }
-        if (self.shadow_mismatches as u64) > config.public_object.max_shadow_mismatches {
+        if self.shadow_mismatches > config.public_object.max_shadow_mismatches {
             blockers.push(AdaptiveReuseBlocker::ShadowMismatch);
         }
         if self.request_count < config.public_object.min_route_samples {
